+++ src/Config/Init/InitializationHandlerConfig.php
@@ -32,7 +32,7 @@ use Ampache\Module\Util\EnvironmentInterface;
 
 final class InitializationHandlerConfig implements InitializationHandlerInterface
 {
-    private const VERSION        = '5.4.0-develop';
+    private const VERSION        = '5.4.0-release';
     private const CONFIG_VERSION = '61';
 
     public const CONFIG_FILE_PATH = __DIR__ . '/../../../config/ampache.cfg.php';