+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@
   * Add `file` to album and artist search
 * CLI
   * New argument for run:updateCatalog `-f|--find` Find missing files and print a list of filenames
+  * New argument for cleanup:sortSongs `-f|--files` Rename files and keep them in the current folder
   * New argument for cleanup:sortSongs `-l|--limit` Limit how many moves to allow before stopping
   * New argument for cleanup:sortSongs `[catalogName]` Name of Catalog (optional)
 * Database 540002: